Created Solver interface, outsourced Status enum
