+++ src/main/java/de/nqueensfaf/SolverStatus.java
@@ -1,0 +1,5 @@
+package de.nqueensfaf;
+
+public enum SolverStatus {
+	IDLE, INITIALIZING, RUNNING, TERMINATING, FINISHED, CANCELED;
+}
